backend/requirements.txt
Django==5.2.1
django-cors-headers==4.7.0
django-environ==0.12.0
djangorestframework==3.16.0
djangorestframework-simplejwt==5.5.0
django-environ==0.12.0
drf-spectacular==0.28.0
drf-spectacular-sidecar==2025.7.1

sqlparse==0.5.3
tzdata==2025.2
setuptools<81

gunicorn==23.0.0
packaging==25.0

psycopg2-binary>=2.9

google-auth-oauthlib==1.2.0
requests-oauthlib==2.0.0
oauthlib==3.3.1